update validation errors

+++ server/controllers/contactController.ts
@@ -1,9 +1,17 @@
 import { Request, Response } from "express";
 import Contact from "../models/contactDetails";
+import { validationResult } from "express-validator";
 
 const createContact = async (req: Request, res: Response) => {
-  const { name, mobileNumber, email, address } = req.body;
-  const contact = new Contact({ name, mobileNumber, email, address });
+  console.log(req.body)
+  const validationErrors = validationResult(req);
+  if (!validationErrors.isEmpty()) {
+    console.log(validationErrors);
+    res.status(404).send(validationErrors);
+    return;
+  }
+  const { name, phone, email, address } = req.body;
+  const contact = new Contact({ name, phone, email, address });
   try {
     const response = await contact.save()
     console.log(`Success: ${response}`);
@@ -26,11 +34,17 @@ const readContacts = async (req: Request, res: Response) => {
 };
 
 const updateContact = async (req: Request, res: Response) => {
-  const { contactId, name, mobileNumber, email, address } = req.body;
+  const validationErrors = validationResult(req);
+  if (!validationErrors.isEmpty()) {
+    console.log(validationErrors);
+    res.status(404).send(validationErrors);
+    return;
+  }
+  const { contactId, name, phone, email, address } = req.body;
   try {
     const response = await Contact.findOneAndUpdate(
       { _id: contactId },
-      { $set: { name: name, mobileNumber: mobileNumber, email: email, address: address } });
+      { $set: { name: name, phone: phone, email: email, address: address } });
     console.log(`Success: ${response}`);
     const contacts = await Contact.find({});
     res.status(200).json(contacts);
@@ -41,6 +55,12 @@ const updateContact = async (req: Request, res: Response) => {
 };
 
 const deleteContact = async (req: Request, res: Response) => {
+  const validationErrors = validationResult(req);
+  if (!validationErrors.isEmpty()) {
+    console.log(validationErrors);
+    res.status(404).send(validationErrors);
+    return;
+  }
   try {
     const response = await Contact.findOneAndDelete({ _id: req.body.contactId });
     console.log(`Success: ${response}`);
